fix/plant detail 모듈화
