+++ src/screens/plant-detail/modules/plant-details-content/index.ts
@@ -1,0 +1,1 @@
+export * from './plant-details-content.module';
