+++ src/screens/plant-detail/modules/plant-details-content/plant-detail-details/index.ts
