+++ src/screens/plant-detail/modules/plant-details-content/plant-detail-details/plant-detail-details.tsx
@@ -69,16 +69,18 @@ export const PlantDetailDetailsModule = memo<PlantDetailDetailsModule>(() => {
 
   return (
     <>
-      <Stack space={4} horizontal>
-        <Icon color={palette['primary']} size={16} name={'list-alt'} />
-        <Text variants={'bodyMedium'} fontWeight={'Medium'} color={'primary'}>
-          상세정보
-        </Text>
-      </Stack>
-      <Stack space={4}>
-        {plantDetails.map((detail, index) => (
-          <DetailItem key={index} detail={detail} />
-        ))}
+      <Stack space={12}>
+        <Stack space={4} horizontal>
+          <Icon color={palette['primary']} size={16} name={'list-alt'} />
+          <Text variants={'bodyMedium'} fontWeight={'Medium'} color={'primary'}>
+            상세정보
+          </Text>
+        </Stack>
+        <Stack space={4}>
+          {plantDetails.map((detail, index) => (
+            <DetailItem key={index} detail={detail} />
+          ))}
+        </Stack>
       </Stack>
     </>
   );
