+++ src/screens/plant-detail/modules/plant-details-content/plant-detail-progress/index.ts
@@ -1,0 +1,1 @@
+export * from './plant-detail-progress.module';
