+++ src/screens/plant-detail/modules/plant-details-content/plant-details-content.module.tsx
@@ -3,10 +3,10 @@ import { useNavigation } from '@react-navigation/native';
 import { memo } from 'react';
 
 import { PlantDetailScreenNavigationProps } from '../../plant-detail.screen';
-import { PlantDetailHeaderModule } from '../plant-detail-header';
 
-import { PlantDetailDetailsModule } from './plant-detail-details/plant-detail-details';
+import { PlantDetailDetailsModule } from './plant-detail-details';
 import { PlantDetailPictureModule } from './plant-detail-picture';
+import { PlantDetailProgressModule } from './plant-detail-progress';
 import { PlantDetailGrowthModule } from './plant-details-growth';
 
 type PlantDetailContentModule = {};
@@ -15,11 +15,15 @@ export const PlantDetailContentModule = memo<PlantDetailContentModule>(() => {
   const navigation = useNavigation<PlantDetailScreenNavigationProps>();
 
   return (
-    <Stack>
-      <PlantDetailHeaderModule />
-      <PlantDetailPictureModule />
-      <PlantDetailDetailsModule />
-      <PlantDetailGrowthModule />
+    <Stack space={12} paddingTop={48}>
+      <Stack space={16}>
+        <PlantDetailPictureModule />
+        <PlantDetailDetailsModule />
+      </Stack>
+      <Stack space={16}>
+        <PlantDetailGrowthModule />
+        <PlantDetailProgressModule />
+      </Stack>
     </Stack>
   );
 });
